requirements.txt
django-crispy-forms
django-allauth
Pillow
django